Rename apiTypes.ts to externalAPITypes.ts
